requirements.txt
coreapi==2.3.3
djangorestframework==3.8.2
Django==2.0.7
factory-boy==2.11.1
feedparser==5.2.1
Markdown==2.6.11
psycopg2-binary==2.7.5
pytest-django==3.2.1
uWSGI==2.0.17
